router: update network matching data structure

+++ envoy/network/filter.h
@@ -219,8 +219,8 @@ public:
 
   virtual OptRef<const Address::Ip> sourceIp() const PURE;
   virtual OptRef<const Address::Ip> destinationIp() const PURE;
-  virtual absl::optional<uint16_t> sourcePort() const PURE;
-  virtual absl::optional<uint16_t> destinationPort() const PURE;
+  virtual absl::optional<uint32_t> sourcePort() const PURE;
+  virtual absl::optional<uint32_t> destinationPort() const PURE;
 };
 
 /**
